refactor: move workflow editor test fixtures
